StudentNotFound and BadRequest

+++ src/main/java/com/example/demo/student/StudentService.java
@@ -1,4 +1,6 @@
 package com.example.demo.student;
+import com.example.demo.student.exception.BadRequestException;
+import com.example.demo.student.exception.StudentNotFoundException;
 
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -16,6 +18,12 @@ public class StudentService {
 
     public void addStudent(Student student) {
         //To do later: check if email is taken
+        Boolean existsEmail = studentRepository
+                .selectExistsEmail(student.getEmail());
+        if(existsEmail) {
+            throw new BadRequestException(
+                    "Email " + student.getEmail() + " taken");
+        }
 
         studentRepository.save(student);
 
@@ -23,6 +31,10 @@ public class StudentService {
 
     public void deleteStudent(Long studentId) {
         //check if student exists
+        if(!studentRepository.existsById(studentId)) {
+            throw new StudentNotFoundException(
+                    "Student with iid " + studentId + " does not exists");
+        }
         studentRepository.deleteById(studentId);
     }
 }
